[#2089]Move to new package. Make all mapper transform() methods static.
